refactor: add svg component to linkBtn
