+++ package.json
@@ -26,7 +26,7 @@
     "eslint-config-next": "13.3.0",
     "i18next": "^22.4.14",
     "i18next-browser-languagedetector": "^7.0.1",
-    "i18next-http-backend": "^2.2.0",
+    "i18next-http-backend": "^2.2.1",
     "next": "13.3.0",
     "next-i18next": "^13.2.2",
     "react": "18.2.0",
