+++ src/ui/Svg/icons.tsx
@@ -35,6 +35,18 @@ export const names = {
   share: (
     <path d='M17.778 5.333v15.556q0 0.444-0.667 0.889t-1.111 0.444-1.111-0.444-0.667-0.889v-15.556l-1.778 2.222q-0.444 0.444-1.111 0.444t-1.111-0.444-0.444-1.111 0.444-1.111l4.444-4.889q0.889-0.444 1.333-0.444t1.333 0.444l4.444 4.889q0.444 0.444 0.444 1.111t-0.444 1.111-1.111 0.444-1.111-0.444zM3.111 16q0.889 0 1.333 0.444t0.444 1.333v8l0.444 1.333q0.444 0.889 1.333 1.333l1.333 0.444h16l1.333-0.444q0.889-0.444 1.333-1.333l0.444-1.333v-8q0-0.889 0.444-1.333t1.111-0.444 1.111 0.444 0.444 1.333v8.444q0 1.333-0.667 2.667t-2.222 2.222-2.889 0.889h-16.889q-1.333 0-2.889-0.889t-2.222-2.222-0.667-2.667v-8.444q0-0.889 0.444-1.333t0.889-0.444z'></path>
   ),
+  appStore: (
+    <path
+      fill='#fff'
+      d='M18.472 6.564c-1.491 1.071-2.237 2.36-2.237 3.867 0 1.805.922 3.19 2.765 4.155-.495 1.454-1.211 2.72-2.15 3.798C15.914 19.46 15.057 20 14.282 20c-.365 0-.863-.122-1.495-.367l-.303-.119c-.618-.244-1.166-.367-1.641-.367-.45 0-.941.096-1.475.288l-.381.139-.48.198c-.377.152-.758.228-1.142.228-.906 0-1.905-.757-3-2.27C2.788 15.56 2 13.197 2 10.64c0-1.818.492-3.283 1.475-4.393.984-1.11 2.286-1.666 3.908-1.666.606 0 1.173.112 1.7.337l.362.149.38.159c.34.145.613.218.822.218.267 0 .563-.063.889-.189l.498-.198.371-.139c.593-.218 1.247-.327 1.964-.327 1.7 0 3.068.658 4.103 1.973zM14.447 0c.02.231.03.41.03.535 0 1.144-.41 2.147-1.232 3.01-.82.863-1.774 1.294-2.862 1.294-.033-.258-.05-.443-.05-.555 0-.972.382-1.884 1.144-2.737C12.239.694 13.122.195 14.125.05c.071-.014.179-.03.322-.05z'
+    ></path>
+  ),
+  googlePlay: (
+    <path
+      fill='#fff'
+      d='M15.3099 12.7188L14.0282 11.386L9.78055 15.9436L15.3099 12.7188Z M16.958 11.7576L17.0071 11.729C18.3303 10.9572 18.3303 9.04538 17.0071 8.27366L16.9564 8.2441L15.3199 10L16.958 11.7576Z M15.3082 7.28288L9.78589 4.06215L14.0282 8.61402L15.3082 7.28288Z M4.04121 1.00037L12.6954 10L4.03865 19.0023C2.98068 19.0214 2 18.1809 2 16.9996V3.00301C2 1.82076 2.98226 0.979819 4.04121 1.00037Z'
+    ></path>
+  ),
   // SVG Extension
   top10: (
     <svg
@@ -64,7 +76,7 @@ export const names = {
       </defs>
     </svg>
   ),
-  "18plus": (
+  '18plus': (
     <svg
       xmlns='http://www.w3.org/2000/svg'
       width='24'
